Basic model upload
Models can be uploaded to the Blobstore.  A record of all models and
related data is stored in the Datastore.  You are required to be signed
into your Google account to upload a model so the user can be recorded
in the Datastore model entry.  Minor file resturcturing so old hello
world example wont be in the way.